Possible fix for MOB-265
git-svn-id: svn+ssh://none.cs.umass.edu/svn/mobility-first/GNS/trunk@815 0b15c3b6-4604-4ca4-ad59-2e1af984e93c

+++ src/edu/umass/cs/gns/nsdesign/recordmap/ReplicaControllerRecord.java
@@ -363,7 +363,7 @@ public class ReplicaControllerRecord {
    */
   public short getActiveVersion() throws FieldNotFoundException {
     if (hashMap.containsKey(ACTIVE_VERSION)) {
-      return (Short) hashMap.get(ACTIVE_VERSION);
+      return ((Integer) hashMap.get(ACTIVE_VERSION)).shortValue();
     }
     throw new FieldNotFoundException(ACTIVE_VERSION);
   }
@@ -375,7 +375,7 @@ public class ReplicaControllerRecord {
    */
   public short getOldActiveVersion() throws FieldNotFoundException {
     if (hashMap.containsKey(OLD_ACTIVE_VERSION)) {
-      return (Short) hashMap.get(OLD_ACTIVE_VERSION);
+      return ((Integer) hashMap.get(OLD_ACTIVE_VERSION)).shortValue();
     }
     throw new FieldNotFoundException(OLD_ACTIVE_VERSION);
   }
